LangChain_projects/requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
langchain
bs4 
pypdf
chromadb
faiss-cpu
wikipedia
arxiv
langchainhub
beautifulsoup4
langchain-groq
sentence-transformers
langchain-objectbox
faiss-cpu
PyPDF2
langchain_google_genai
google-cloud-aiplatform